added get_#{association_name} helper for (many/one)_to_one associations.
This is mainly to support common usecases in fields_for. This method
will return a new record if the associated object is nil at the time
it is accessed, and the associated object otherwise. This should help
fields_for to render the right id and name attributes for the input
fields. This method is not generated for (one/many)_to_many associations
since these return an Collection anyway and this allow to call build
or get on it. A possible use case would look something like this:

<% fields_for @location.get_geo_point do %>
 ...
<% end =%>

The whole purpose of this commit is to provide something to fields_for
that is guaranteed to not be nil.