Completely removed transactional save behavior
All specs pass on all official adapters on all
supported rubies. I also tested those configurations
with either dm-validations and/or dm-constraints
enabled and the specs pass all these permutations
too. Instructions on how to run specs using bundle
exec can be found in the Gemfile.

It's not feasible for a datamapper plugin of this
kind to force a hard dependency on dm-transactions.
Transactions are not available on every adapter,
so depending on dm-transactions would make this
plugin useless for those adapters, which is neither
desired or necessary.

In general, it's safer to put the control over
which transactions get executed directly into the
hands of the user. That might be a bit more
inconvenient sometimes, but really the only safe
decision to make.

If you've previously been relying on transaction
behavior (when updating associated records that
fail validation for example), make sure that your
client code is wrapped inside a transaction
explicitly.

Also, the (completely broken) code for validation
error collection has been removed with this commit.
A working version will be added soon though.