Support for composite keys.
The resource and collection setters for nested attributes were expecting
single-column :id primary keys. This breaks if the primary key has a
different name or is a composite key.

This patch removes the hard-coded :id value and instead extracts the
proper values for a (composite) primary key by looking at the models
involved in the association. Properties of the primary key that are
included in the foreign key of the association are taken from the model
instance, any missing primary key properties have to be specified in the
attribute hashes passed to the nested attribute setters.

Further it changes the logic of the setters so that we always check for
existing records if we have a full key because having these values does
not imply existence (e.g. for n:m join tables using a composite primary
key, having the foreign key values to both joined entities says nothing
about the actual existence of such a record).