Split unassignable keys into keys used for update and create.
Primary key values should be excluded when updating a resource. With nested
attribute setters it is impossible anyway to change primary key values. But
for creation the primary key values cannot be excluded because not every
primary key property is a Serial (e.g. n:m join tables with a composite
primary key).